refactor(assistant): relocate DB core into persistence/ with re-export shims
Move the DB connection/init/maintenance/singleton/async-query modules and the
schema/ tree from memory/ into a new persistence/ layer. The old memory/ paths
become re-export shims so every existing importer keeps working unchanged;
importer migration and shim removal come in later PRs.

Part of the memory-as-a-plugin persistence split (PR 1 of 8).
