+++ assistant/src/persistence/schema/a2a.ts
