+++ assistant/src/persistence/schema/bookmarks.ts
