+++ assistant/src/persistence/schema/contacts.ts
