+++ assistant/src/persistence/schema/index.ts
