+++ assistant/src/persistence/schema/infrastructure.ts
