+++ assistant/src/persistence/schema/memory-core.ts
