+++ assistant/src/persistence/schema/memory-graph.ts
