+++ assistant/src/persistence/schema/tasks.ts
